+++ superset-frontend/src/components/index.ts
@@ -49,9 +49,3 @@ export {
   type PluginContextType,
 } from './DynamicPlugins';
 export * from './FacePile';
-export {
-  GlossaryTooltip,
-  GLOSSARY_TERMS,
-  type GlossaryTooltipProps,
-  type GlossaryTerm,
-} from './GlossaryTooltip';
